hackathon/chatbot/requirements.txt
streamlit
pandas
langchain-core
langchain-openai
python-dotenv
./actiancortex-0.1.0b1-py3-none-any.whl
